+++ NEWS.md
@@ -6,12 +6,14 @@
 -   New function `simpleStack`, a simple stack data structure with methods in R6 class. Methods include `pop`, `push` and more.
 -   New function `historyStack`, a stack data structure to store history steps in R6 class. Similar to browser, word editor or image editor history, stored data can be moved forward or backward by step(s).
 -   New in-line operation functions `inc`, `mult`, `divi`, which are equivalent of `1 += 1`, `1 *= 2`, `i /= 2` in other programming languages.
+-   New function `timeout`, run expressions with time limits.
+-   `checkNameSpace` has new arguments, `time_out` and `on_timeout`. This handles if the installed package list to large and waiting time is too long. Instead of waiting, when times out, directly return all checking packages as missing, but can be customize to other behavior as well.
 
 ## Minor Change
 
 -   `emptyIsFalse` now named `notFalsy` and its opposite is `isFalsy`, old name `emptyIsFalse` still reminds in the package.
 
-## bug fix
+## Bug fix
 
 -   Fix NA bugs in `emptyIsFalse`.